Fix menu on small screens
On small screens (600px or smaller right now) the menu pops up after clicking the hamburger icon. But, the menu items would move the rest of the content down. I fixed that so the content below the menu stays put as the menu toggles displaying and not. This involved adding a #menu div in the index.html file to contain the menu items.